Don't run curl as root
There's no particularly good reason to run the curl as root, so move that out of the sudo block.

Also switch the `-` to `:-` in script location; the former only applies if the variable is undefined, while the latter also applies if the variable is defined but empty.  And delete the vestigial `()` after the function name while I'm in there.  :)